+++ index.js
@@ -1,0 +1,226 @@
+const os = require('os')
+const path = require('path')
+
+const Hyperswarm = require('hyperswarm')
+const IdEnc = require('hypercore-id-encoding')
+const Corestore = require('corestore')
+const HyperDHT = require('hyperdht')
+const rrp = require('resolve-reject-promise')
+const safetyCatch = require('safety-catch')
+const Hyperdrive = require('hyperdrive')
+const pino = require('pino')
+
+const BlindPeerClient = require('@holepunchto/blind-peering')
+const LookupClient = require('autobase-discovery/client/lookup.js')
+const goodbye = require('graceful-goodbye')
+
+const DEFAULT_MIN = 3
+const DEFAULT_LIMIT = 6
+const DEFAULT_STORAGE = path.join(path.normalize(os.homedir()), '.blind-peering-cli-storage')
+const DEFAULT_TIMEOUT_SEC = 15
+
+async function seed (opts) {
+  try {
+    const key = IdEnc.decode(opts.key)
+    const { autoDiscDb, serviceName, debug } = opts
+    const blindPeerKey = opts.blindPeerKey ? IdEnc.decode(opts.blindPeerKey) : null
+    if ((blindPeerKey && autoDiscDb) || (!blindPeerKey && !autoDiscDb)) {
+      throw new Error('Must set exactly one of blind-peer-key and auto-disc-db')
+    }
+
+    const logger = pino({
+      level: 'info',
+      transport: {
+        target: 'pino-pretty'
+      }
+    })
+
+    const min = blindPeerKey
+      ? 1
+      : parseInt(opts.min || DEFAULT_MIN)
+
+    const limit = blindPeerKey
+      ? 1
+      : parseInt(opts.limit || DEFAULT_LIMIT)
+    const msTimeout = parseInt(opts.timeout || DEFAULT_TIMEOUT_SEC) * 1000
+    const storage = path.normalize(opts.storage || DEFAULT_STORAGE)
+
+    logger.info(`Using storage ${storage}`)
+    const { store, swarm } = await getStoreAndSwarm(storage)
+    logger.info(`Using DHT public key: ${IdEnc.normalize(swarm.dht.defaultKeyPair.publicKey)}`)
+
+    swarm.on('connection', (conn, peerInfo) => {
+      if (debug) {
+        const key = IdEnc.normalize(peerInfo.publicKey)
+        logger.debug(`Opened connection to ${key}`)
+        conn.on('close', () => logger.debug(`Closed connection to ${key}`))
+      }
+      store.replicate(conn)
+    })
+
+    let done = false
+    let seedTimeout = null
+    let client = null
+    goodbye(async () => {
+      clearTimeout(seedTimeout)
+      logger.info(done ? 'Shutting down...' : 'Cancelling...')
+      await swarm.destroy()
+      if (client) await client.close()
+      await store.close()
+    })
+
+    const blindPeers = []
+    if (blindPeerKey) {
+      blindPeers.push(blindPeerKey)
+    } else {
+      logger.info(`Requesting up to ${limit} instances from the '${serviceName}' service, using autobase-discovery database ${IdEnc.normalize(autoDiscDb)}`)
+      client = new LookupClient(
+        autoDiscDb, swarm, store.namespace('autodiscovery-lookup')
+      )
+
+      await client.ready()
+
+      for await (const { publicKey } of await client.list(serviceName, { limit })) {
+        blindPeers.push(publicKey)
+      }
+      logger.info(`Using blind-peers:\n  -${(blindPeers.map(p => IdEnc.normalize(p)).join('\n  -'))}`)
+
+      if (blindPeers.length < min) {
+        logger.error(`Found only ${blindPeers.length} peers, whereas a minimum of ${min} was specified`)
+        process.exit(1)
+      }
+    }
+
+    // TODO: should ideally live in the blind peer (detecting when it's a hyperdrive)
+    logger.info('Obtaining the blobs key...')
+    const [dbCore, blobsCore] = await getDbAndBlobs(store, key, swarm)
+    const blobsKey = blobsCore.key
+
+    logger.info(`Requesting ${blindPeers.length} blind peers to seed core ${IdEnc.normalize(key)} and blobs core ${IdEnc.normalize(blobsKey)} (minimum successes: ${min})`)
+
+    const blindPeerClient = new BlindPeerClient(swarm, store, { coreMirrors: blindPeers })
+    const seedProms = []
+    for (const b of blindPeers) {
+      seedProms.push(
+        Promise.all([
+          blindPeerClient.addCore(dbCore.session(), b, { announce: true }),
+          blindPeerClient.addCore(blobsCore.session(), b, { announce: true })
+        ])
+      )
+    }
+
+    const { resolve: resolveSeed, reject: rejectSeed, promise: seedProm } = rrp()
+    seedTimeout = setTimeout(
+      () => rejectSeed(new Error(`Timeout: failed to request ${min} blind peers to seed the core`)),
+      msTimeout
+    )
+
+    let nrResolved = 0
+    let nrRejected = 0
+    const processReject = () => {
+      nrRejected++
+      if (nrRejected > blindPeers.length - min) rejectSeed(new Error(`Failed to request seeding at least ${min} blind peers`))
+    }
+
+    for (let i = 0; i < seedProms.length; i++) {
+      const prom = seedProms[i]
+      const normBlindPeerKey = IdEnc.normalize(blindPeers[i])
+
+      prom.then(
+        ([dbRes, blobsRes]) => {
+          if (!dbRes) {
+            processReject()
+            logger.warn(`Could not request db core from blind peer ${normBlindPeerKey}`)
+            return
+          }
+          if (dbRes.announce !== true) {
+            logger.warn(`You are not a trusted peer for blind peer ${normBlindPeerKey} (announce was downgraded to false)`)
+            processReject()
+            return
+          }
+          if (!blobsRes) {
+            logger.warn(`Could not request blobs core from blind peer ${normBlindPeerKey}`)
+            processReject()
+            return
+          }
+
+          nrResolved++
+          logger.info(`Successfully contacted seeder ${normBlindPeerKey} (successes: ${nrResolved}, failures: ${nrRejected})`)
+          if (nrResolved >= min) {
+            resolveSeed()
+          }
+        },
+        (err) => {
+          safetyCatch(err)
+          if (!done) logger.info(`Failed to contact seeder ${normBlindPeerKey} (successes: ${nrResolved}, failures: ${nrRejected})`)
+          if (debug) logger.warn(`Error while communicating with seeder ${normBlindPeerKey}: ${err.stack}`)
+          processReject()
+        }
+      )
+    }
+
+    try {
+      await seedProm
+    } catch (e) {
+      if (debug) {
+        logger.error(e)
+      } else {
+        logger.error(e.message)
+      }
+      goodbye.exit() // TODO: a way to signal non-zero exit code
+      return
+    }
+    done = true
+
+    // TODO: verify it was announced + verify it is downloaded
+    logger.info(`Successfully requested to seed hyperdrive ${IdEnc.normalize(key)} with blobs core ${IdEnc.normalize(blobsKey)}`)
+    goodbye.exit()
+  } catch (e) {
+    console.error(e.message)
+    goodbye.exit() // TODO: a way to signal non-zero exit code
+  }
+}
+
+async function identity (opts) {
+  const logger = console
+
+  const storage = path.normalize(opts.storage || DEFAULT_STORAGE)
+
+  logger.info(`Using storage ${storage}`)
+  const { store, swarm } = await getStoreAndSwarm(storage)
+  const ownKey = IdEnc.normalize(swarm.dht.defaultKeyPair.publicKey)
+  logger.info(`Your DHT public key is: ${ownKey}`)
+  logger.info(`To be able to send 'seed' requests to blind peers, ask them to add the '--trusted-peer ${ownKey}' flag when launching their blind peer instance`)
+
+  await swarm.destroy()
+  await store.close()
+}
+
+async function getDbAndBlobs (store, key, swarm) {
+  const drive = new Hyperdrive(store.namespace('drive'), key)
+  await drive.ready()
+  await new Promise(resolve => {
+    if (drive.blobs) resolve()
+    else {
+      drive.once('blobs', resolve)
+      swarm.join(drive.discoveryKey)
+      drive.getBlobs().catch(safetyCatch)
+    }
+  })
+
+  return [drive.db.core, drive.blobs.core]
+}
+
+async function getStoreAndSwarm (storage) {
+  const store = new Corestore(storage)
+  await store.ready()
+
+  // We need a consistent keypair across restarts, because we use
+  // an allow-list at the blind-peer side (to which our key should be added)
+  const keyPair = await store.createKeyPair('dht-client-identity')
+  const swarm = new Hyperswarm({ dht: new HyperDHT({ keyPair }) })
+
+  return { store, swarm }
+}
+
+module.exports = { seed, identity }
