+++ package.json
@@ -2,7 +2,7 @@
   "name": "@holepunchto/blind-peering-cli",
   "version": "1.1.0",
   "description": "CLI for blind-peering",
-  "main": "bin.js",
+  "main": "index.js",
   "scripts": {
     "test": "standard"
   },
